pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = ""
authors = ["Jay"]
readme = "README.md"
packages = [{include = "ai_agent"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
openai = "^1.3.7"
mysql-connector-python = "^8.2.0"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
pyautogen = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "ai_agent.main:main"
